Implement static methods for creating BIDS file objects from BIDS dataset paths
