+++ src/bids/types/tsv.js
@@ -54,6 +54,32 @@ export class BidsTsvFile extends BidsFile {
     this._parseHedColumn()
   }
 
+  /**
+   * Parse a BIDS TSV file from a BIDS dataset path and sidecar.
+   *
+   * @param {string} datasetRoot The root path of the dataset.
+   * @param {string} relativePath The relative path of the file within the dataset.
+   * @param {BidsSidecar} sidecar The BIDS sidecar to use with this file.
+   * @returns {Promise<BidsTsvFile>} The built TSV file object.
+   */
+  static async createFromBidsDatasetPathAndSidecar(datasetRoot, relativePath, sidecar) {
+    const jsonData = sidecar.jsonData
+    return BidsTsvFile.createFromBidsDatasetPathAndJson(datasetRoot, relativePath, jsonData)
+  }
+
+  /**
+   * Parse a BIDS TSV file from a BIDS dataset path and sidecar JSON data.
+   *
+   * @param {string} datasetRoot The root path of the dataset.
+   * @param {string} relativePath The relative path of the file within the dataset.
+   * @param {object} jsonData The BIDS sidecar data to use with this file.
+   * @returns {Promise<BidsTsvFile>} The built TSV file object.
+   */
+  static async createFromBidsDatasetPathAndJson(datasetRoot, relativePath, jsonData) {
+    const [contents, fileObject] = await BidsFile.readBidsFileFromDatasetPath(datasetRoot, relativePath)
+    return new BidsTsvFile(relativePath, fileObject, contents, jsonData)
+  }
+
   _parseHedColumn() {
     const hedColumn = this.parsedTsv.get('HED')
     if (hedColumn === undefined) {
